requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20241021
Django==5.2.1
django-auth-ldap==5.2.0
django-cors-headers==4.7.0
django-filter==25.1
django-impersonate==1.9.5
django-import-export==4.3.7
django-rosetta==0.10.2
django-translations==1.4.4
django_celery_results==2.6.0
djangorestframework==3.16.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
polib==1.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ldap==3.4.4
pytz==2025.2
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tornado==6.5.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
